move promptlab to its own directory
