+++ tests/promptlab/test_promptlab_model.py
@@ -2,9 +2,9 @@ from unittest import mock
 
 import pandas as pd
 
-from mlflow._promptlab import _PromptlabModel
 from mlflow.entities.param import Param
 from mlflow.gateway import set_gateway_uri
+from mlflow.promptlab import _PromptlabModel
 
 set_gateway_uri("http://localhost:5000")
